+++ CHANGES.md
@@ -3,7 +3,7 @@
 # v 1.37
 
 * Changes
-  * Add client side raster reprojection.
+  * Add client side raster reprojection, see the example to see how to enable it.
 
 # v 1.36 - 2018-04-18